Add Screenshot Utilities KLOV Provider configuration and update my other addons. (#504)
* Add Screenshot Utilities KLOV Provider configuration

New provider addon for Screenshot Utilities that fetches screenshots from the Killer List Of Video Games.

* Update ShortDescription for KLOVProvider addon

* Add help and donation links to YAML

Added additional links for help and donations.

* Add Help and Ko-fi links to HerrKnarz_LinkUtilities

* Add help and donation links and added features to description

* Update description to indicate addon deprecation

This addon is deprecated and will only receive bugfixes. The features are merged into and extended in Metadata Utilities.

* Enhance description for GOGProvider addon

Updated the description to clarify functionality regarding fetching screenshots.

* Enhance description for Screenshot Utilities addon

Updated the description to clarify functionality regarding screenshot fetching and the search function.

* Refine description for Screenshot Utilities addon

Updated the description to clarify functionality and usage.